improve markup: scale the image for the edits to 800px width
The annotation settings of excalidraw are optimized for a canvas size of up to 800 px, larger canvases seem to clip arrow heads. This PR scales the images (not the source!) to a width of 800px. The export keeps the original size with a max width of 3840px (higher does not make sense on web).